refactor(components/molecule/badgeCounter): api tab
